Fix fp alignment
Commit a5bbbd8a324a fixed alignment earlier up, but then screwed
it up further down. Fix it.

+++ stat.h
@@ -181,7 +181,7 @@ struct thread_stat {
 
 	uint32_t latency_depth;
 	uint64_t latency_target;
-	uint16_t pad;
+	uint32_t pad;
 	fio_fp64_t latency_percentile;
 	uint64_t latency_window;
 } __attribute__((packed));
